chore(django-allauth): enable github sign in

+++ src/registry/api.py
@@ -1,10 +1,22 @@
 import datetime
-from typing import List, Optional
+from typing import List
+
+from ninja import Query, Router
+from ninja.orm.fields import AnyObject
+from ninja.security import django_auth
+
+from main.models import CustomUser
+from main.schema import ForbiddenSchema, NotFoundSchema, Schema, SuccessSchema
 
 from .models import Author, Model, Prediction
-from ninja import Router
-from ninja.orm.fields import AnyObject
-from .schema import AuthorSchema, ModelSchema, NotFoundSchema, PredictionSchema, Schema
+from .schema import (
+    AuthorFilterSchema,
+    AuthorSchema,
+    ModelFilterSchema,
+    ModelSchema,
+    PredictionSchema,
+    PredictionFilterSchema,
+)
 
 router = Router()
 
@@ -13,124 +25,196 @@ router = Router()
 class AuthorIn(Schema):
     """Input for the request's body"""
 
-    name: str
-    email: str
+    user: str
+    institution: str
+
+
+class AuthorInPost(Schema):
+    """Input for POST update request's body"""
+
     institution: str
 
 
 @router.get("/authors/", response=List[AuthorSchema])
-def list_authors(request, name: Optional[str] = None):
+def list_authors(
+    request,
+    filters: AuthorFilterSchema = Query(...),
+):
     """Lists all authors, can be filtered by name"""
-    if name:
-        return Author.objects.filter(name__icontains=name)
-    return Author.objects.all()
+    authors = Author.objects.all()
+    authors = filters.filter(authors)
+    return authors
 
 
-@router.get("/authors/{author_id}", response={200: AuthorSchema, 404: NotFoundSchema})
-def get_author(request, author_id: int):
-    """Gets author by id"""
+@router.get("/authors/{username}", response={200: AuthorSchema, 404: NotFoundSchema})
+def get_author(request, username: str):
+    """Gets author by Github username"""
     try:
-        author = Author.objects.get(pk=author_id)
-        return (200, author)
+        author = Author.objects.get(user__username=username)
+        return 200, author
     except Author.DoesNotExist:
-        return (404, {"message": "Author not found"})
+        return 404, {"message": "Author not found"}
 
 
-@router.post("/authors/", response={201: AuthorSchema})
+@router.post(
+    "/authors/",
+    response={201: AuthorSchema, 403: ForbiddenSchema, 404: NotFoundSchema},
+    auth=django_auth,
+)
 def create_author(request, payload: AuthorIn):
-    """Posts author to database"""
-    author = Author.objects.create(**payload.dict())
-    return (201, author)
-
-
-@router.put("/authors/{author_id}", response={201: AuthorSchema, 404: NotFoundSchema})
-def update_author(request, author_id: int, payload: AuthorIn):
-    """Updates author"""
+    """Posts author to database, requires a CustomUser to be created"""
     try:
-        author = Author.objects.get(pk=author_id)
+        user = CustomUser.objects.get(username=payload.user)
+        try:
+            author = Author.objects.get(user__username=payload.user)
+            if author:
+                return 403, {"message": f"Author '{author}' already registered"}
+        except Author.DoesNotExist:
+            author = Author.objects.create(user=user, institution=payload.institution)
+            return 201, author
+    except CustomUser.DoesNotExist:
+        return 404, {"message": f"User '{payload.user}' does not exist"}
 
-        for attr, value in payload.dict().items():
-            setattr(author, attr, value)
 
+@router.put(
+    "/authors/{username}",
+    response={201: AuthorSchema, 403: ForbiddenSchema, 404: NotFoundSchema},
+    auth=django_auth,
+)
+def update_author(request, username: str, payload: AuthorInPost):
+    """
+    Updates author. It is not possible to change Author's
+    user and this post method can only be called by the user
+    """
+    try:
+        author = Author.objects.get(user__username=username)
+
+        if request.user != author.user:  # TODO: Enable admins here
+            return 403, {"message": "You are not authorized to update this author"}
+
+        author.institution = payload.institution
         author.save()
-        return (201, author)
+        return 201, author
     except Author.DoesNotExist:
-        return (404, {"message": "Author not found"})
+        return 404, {"message": "Author not found"}
 
 
-@router.delete("/authors/{author_id}", response={204: None, 404: NotFoundSchema})
-def delete_author(request, author_id: int):
+@router.delete(
+    "/authors/{username}",
+    response={200: SuccessSchema, 403: ForbiddenSchema, 404: NotFoundSchema},
+    auth=django_auth,
+)
+def delete_author(request, username: str):
     """Deletes author"""
     try:
-        author = Author.objects.get(pk=author_id)
+        author = Author.objects.get(user__username=username)
+
+        if request.user != author.user:  # TODO: Enable admins here
+            return 403, {"message": "You are not authorized to delete this author"}
+
         author.delete()
-        return 204
+        return 200, {"message": f"Author '{author.user.name}' deleted successfully"}
     except Author.DoesNotExist:
-        return (404, {"message": "Author not found"})
+        return 404, {"message": "Author not found"}
 
 
 # [Model] Model
 class ModelIn(Schema):
     name: str
     description: str = None
-    author: AuthorSchema
-    repository: str
+    author: str  # Author username
+    repository: str  # TODO: Validate repository?
     implementation_language: str
     type: str
 
 
 @router.get("/models/", response=List[ModelSchema])
-def list_models(request, name: Optional[str] = None):
-    if name:
-        name_filter = Model.objects.filter(name__icontains=name)
-        return list(name_filter.select_related("author"))
-
+def list_models(request, filters: ModelFilterSchema = Query(...)):
     models = Model.objects.all()
-    return models.select_related("author")
+    models = filters.filter(models)
+    return models
 
 
 @router.get("/models/{model_id}", response={200: ModelSchema, 404: NotFoundSchema})
 def get_model(request, model_id: int):
     try:
-        model = Model.objects.get(pk=model_id)
-        return (200, model)
+        model = Model.objects.get(pk=model_id)  # TODO: get model by id?
+        return 200, model
     except Model.DoesNotExist:
-        return (404, {"message": "Model not found"})
+        return 404, {"message": "Model not found"}
 
 
-@router.post("/models/", response={201: ModelSchema})
+@router.post(
+    "/models/",
+    response={201: ModelSchema, 403: ForbiddenSchema, 404: NotFoundSchema},
+    auth=django_auth,
+)
 def create_model(request, payload: ModelIn):
+    try:
+        author = Author.objects.get(user__username=payload.author)
+        if request.user != author.user:
+            return 403, {
+                "message": "You are not authorized to add a Model to this author"
+            }
+        payload.author = author
+    except Author.DoesNotExist:
+        return 404, {"message": "Invalid Author"}
     model = Model.objects.create(**payload.dict())
-    return (201, model)
+    return 201, model
 
 
-@router.put("/models/{model_id}", response={201: ModelSchema, 404: NotFoundSchema})
+@router.put(
+    "/models/{model_id}",
+    response={201: ModelSchema, 403: ForbiddenSchema, 404: NotFoundSchema},
+    auth=django_auth,
+)
 def update_model(request, model_id: int, payload: ModelIn):
     try:
-        model = Model.objects.get(pk=model_id)
+        model = Model.objects.get(pk=model_id)  # TODO: Update by id?
 
-        for attr, value in payload.dict().items():
-            setattr(model, attr, value)
+        if request.user != model.author.user:  # TODO: allow admins here
+            return 403, {"message": "You are not authorized to update this Model"}
 
-        model.save()
-        return (201, model)
+        try:
+            author = Author.objects.get(user__username=payload.author)
+            payload.author = author
+
+            for attr, value in payload.dict().items():
+                setattr(model, attr, value)
+
+            model.save()
+            return 201, model
+        except Author.DoesNotExist:
+            return 404, {
+                "message": (
+                    f"Author '{payload.author}' not found, use username instead"
+                )
+            }
     except Model.DoesNotExist:
-        return (404, {"message": "Model not found"})
+        return 404, {"message": "Model not found"}
 
 
-@router.delete("/models/{model_id}", response={204: None, 404: NotFoundSchema})
+@router.delete(
+    "/models/{model_id}",
+    response={204: SuccessSchema, 403: ForbiddenSchema, 404: NotFoundSchema},
+    auth=django_auth,
+)
 def delete_model(request, model_id: int):
     try:
         model = Model.objects.get(pk=model_id)
+
+        if request.user != model.author.user:
+            return 403, {"message": "You are not authorized to delete this Model"}
+
         model.delete()
-        return 204
+        return 204, {"message": f"Model {model.name} deleted successfully"}
     except Author.DoesNotExist:
-        return (404, {"message": "Model not found"})
+        return 404, {"message": "Model not found"}
 
 
 # [Model] Prediction
 class PredictionIn(Schema):
-    model: ModelSchema
+    model: ModelSchema  # TODO: change it. Issue #20
     description: str = None
     commit: str
     predict_date: datetime.date
@@ -138,13 +222,13 @@ class PredictionIn(Schema):
 
 
 @router.get("/predictions/", response=List[PredictionSchema])
-def list_predictions(request, model_name: Optional[str] = None):
-    if model_name:
-        name_filter = Prediction.objects.filter(model__name__icontains=model_name)
-        return list(name_filter.select_related("model"))
-
-    prediction = Prediction.objects.all()
-    return prediction.select_related("model")
+def list_predictions(
+    request,
+    filters: PredictionFilterSchema = Query(...),
+):
+    predictions = Prediction.objects.all()
+    predictions = filters.filter(predictions)
+    return predictions
 
 
 @router.get(
@@ -153,40 +237,68 @@ def list_predictions(request, model_name: Optional[str] = None):
 )
 def get_prediction(request, predict_id: int):
     try:
-        prediction = Prediction.objects.get(pk=predict_id)
-        return (200, prediction)
+        prediction = Prediction.objects.get(pk=predict_id)  # TODO: get by id?
+        return 200, prediction
     except Prediction.DoesNotExist:
-        return (404, {"message": "Prediction not found"})
+        return 404, {"message": "Prediction not found"}
 
 
-@router.post("/predictions/", response={201: PredictionSchema})
+@router.post(
+    "/predictions/",
+    response={201: PredictionSchema, 403: ForbiddenSchema, 404: NotFoundSchema},
+    auth=django_auth,
+)
 def create_prediction(request, payload: PredictionIn):
-    prediction = Prediction.objects.create(**payload.dict())
-    return (201, prediction)
+    try:
+        model = Model.objects.get(name=payload.model)  # TODO: get by name?
+
+        if request.user != model.author.user:
+            return 403, {
+                "message": "You are not authorized to add a prediction to this model"
+            }
+
+        payload.model = model
+        # TODO: Add commit verification here #19
+        prediction = Prediction.objects.create(**payload.dict())
+        return 201, prediction
+    except Model.DoesNotExist:
+        return 404, {"message": f"Model '{payload.model}' not found"}
 
 
 @router.put(
     "/predictions/{predict_id}",
-    response={201: PredictionSchema, 404: NotFoundSchema},
+    response={201: PredictionSchema, 403: ForbiddenSchema, 404: NotFoundSchema},
+    auth=django_auth,
 )
 def update_prediction(request, predict_id: int, payload: PredictionIn):
     try:
         prediction = Prediction.objects.get(pk=predict_id)
 
+        if request.user != prediction.model.author.user:
+            return 403, {"message": "You are not authorized to update this prediction"}
+
         for attr, value in payload.dict().items():
             setattr(prediction, attr, value)
-
+        # TODO: Add commit verification if commit has changed
         prediction.save()
-        return (201, prediction)
+        return 201, prediction
     except Prediction.DoesNotExist:
-        return (404, {"message": "Prediction not found"})
+        return 404, {"message": "Prediction not found"}
 
 
-@router.delete("/predictions/{predict_id}", response={204: None, 404: NotFoundSchema})
+@router.delete(
+    "/predictions/{predict_id}",
+    response={204: SuccessSchema, 403: ForbiddenSchema, 404: NotFoundSchema},
+    auth=django_auth,
+)
 def delete_prediction(request, predict_id: int):
     try:
         prediction = Prediction.objects.get(pk=predict_id)
+
+        if request.user != prediction.model.author.user:
+            return 403, {"message": "You are not authorized to delete this prediction"}
+
         prediction.delete()
-        return 204
+        return 204, {"message": f"Prediction {prediction.id} deleted successfully"}
     except Prediction.DoesNotExist:
-        return (404, {"message": "Prediction not found"})
+        return 404, {"message": "Prediction not found"}
